examples/transformers/language-generation/requirements.txt
intel_extension_for_pytorch
tqdm
transformers